getRzRyAngバグとり。ついに完成した！（今頃）
git-svn-id: file:///home/gecchi/repos@2295 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/include/jp/ggaf/dx9core/util/GgafDx9Quaternion.h
@@ -11,10 +11,10 @@ namespace GgafDx9Core {
  */
 class GgafDx9Quaternion : public GgafCore::GgafObject {
 public:
-    float   _t; //‹••”
-    float   _x;
-    float   _y;
-    float   _z;
+    double   _t; //‹••”
+    double   _x;
+    double   _y;
+    double   _z;
 
     GgafDx9Quaternion() {
         _t = 0;
@@ -23,15 +23,15 @@ public:
         _z = 0;
     }
 
-    GgafDx9Quaternion(float prm_t, float prm_x, float prm_y, float prm_z);
-    void set(float prm_t, float prm_x, float prm_y, float prm_z) {
+    GgafDx9Quaternion(double prm_t, double prm_x, double prm_y, double prm_z);
+    void set(double prm_t, double prm_x, double prm_y, double prm_z) {
         _t = prm_t;
         _x = prm_x;
         _y = prm_y;
         _z = prm_z;
     }
-    void mul(float t2, float x2, float y2, float z2) {
-        static float t1, x1, y1, z1;
+    void mul(double t2, double x2, double y2, double z2) {
+        static double t1, x1, y1, z1;
         t1 = _t;
         x1 = _x;
         y1 = _y;
